add markdown comment editor
