+++ package.json
@@ -14,6 +14,7 @@
     "*.{js,cjs,mjs,ts,tsx,md,json}": "prettier --write"
   },
   "dependencies": {
+    "@uiw/react-md-editor": "^3.20.6",
     "emoji-picker-react": "^4.4.8",
     "marked-react": "^1.3.2",
     "moment": "^2.29.4",
